Add two options to mergeChannels
mergeChannels can be used to "div" one channel from another with a
constant slope. That is needed when most of the pixel have no signal
and one want to remove bleedthrough from fluorophore, not from
background.
mergeChannels can also write signed tiff as output. Not sure it is
needed but it is probably safer in case the division subtract a bit
too much.

Update changelog with these and previous changes